third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-element-change-select-attribute-after-deleted-crash-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: host1.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL content-element-change-select-attribute-after-deleted-crash Uncaught TypeError: host1.createShadowRoot is not a function
Harness: the test ran to completion.
